refactor: move url constants
