--- frontend/src/api/chat/constants.ts
@@ -1,1 +1,0 @@
-export const CHAT_URL = '/api/chat';
