+++ frontend/src/api/constants.ts
@@ -1,0 +1,2 @@
+export const CHAT_URL = '/api/chat';
+export const SOURCE_URL = '/api/document';
